+++ crates/derive/src/koto_impl.rs
@@ -1,9 +1,10 @@
-use crate::{PREFIX_FUNCTION, PREFIX_STATIC};
+use crate::PREFIX_FUNCTION;
 use proc_macro::TokenStream;
 use quote::{format_ident, quote};
 use syn::{
-    meta::ParseNestedMeta, parse::Result, parse_macro_input, parse_quote, Attribute, FnArg, Ident,
-    ImplItem, ItemImpl, LitStr, Meta, Path, ReturnType, Signature, Type,
+    meta::ParseNestedMeta, parse::Result, parse_macro_input, parse_quote, Attribute, FnArg,
+    Generics, Ident, ImplItem, ImplItemFn, ItemImpl, LitStr, Meta, Path, ReturnType, Signature,
+    Type, TypePath,
 };
 
 struct KotoImplParser {
@@ -29,28 +30,32 @@ impl KotoImplParser {
     }
 }
 
-pub(crate) fn generate_koto_access_entries(args: TokenStream, item: TokenStream) -> TokenStream {
+// Derives an implementation of `KotoEntries` for types tagged with `#[koto_impl]`
+pub(crate) fn koto_impl(args: TokenStream, item: TokenStream) -> TokenStream {
     let mut attrs = KotoImplParser::default();
     let parser = syn::meta::parser(|meta| attrs.parse(meta));
     parse_macro_input!(args with parser);
     let runtime = attrs.runtime_path;
 
-    let item_clone = item.clone();
-    let input = parse_macro_input!(item_clone as ItemImpl);
+    // Parse the tagged impl block, mutable so that generated functions can be appended.
+    let mut input_struct = parse_macro_input!(item as ItemImpl);
 
-    let struct_ident = match input.self_ty.as_ref() {
-        Type::Path(type_path) => type_path.path.get_ident().expect("Simple ident"),
-        _ => panic!("Expected a struct"),
+    let struct_type = input_struct.self_ty.as_ref();
+    let struct_ident = match &struct_type {
+        Type::Path(TypePath { path, .. }) => {
+            let last_segment = path.segments.last().expect("Expected an identifier");
+            &last_segment.ident
+        }
+        _ => panic!("Expected a type path"),
     };
-    let entries_map_name = format_ident!(
-        "{PREFIX_STATIC}_ENTRIES_{}",
-        struct_ident.to_string().to_uppercase()
-    );
 
-    let (wrapper_functions, insert_ops): (Vec<_>, Vec<_>) = input
+    // Generate wrapper functions an entry map insertion ops for each impl function tagged with
+    // `#[koto_method]`
+    let mut insert_op_count = 0;
+    let (wrapper_functions, insert_ops): (Vec<_>, Vec<_>) = input_struct
         .items
         .iter()
-        // find impl funtions tagged with #[koto_method]
+        // Find impl funtions tagged with #[koto_method]
         .filter_map(|item| match item {
             ImplItem::Fn(f) => f
                 .attrs
@@ -59,180 +64,158 @@ pub(crate) fn generate_koto_access_entries(args: TokenStream, item: TokenStream)
                 .map(|attr| (f, attr)),
             _ => None,
         })
-        // Generate wrappers and lookup inserts for each koto method
-        .map(|(f, attr)| {
-            let (wrapper, wrapper_name) = wrap_function(&f.sig, struct_ident, &runtime);
-            let insert = insert_wrapped_function(&f.sig, attr, wrapper_name, &runtime);
-            (wrapper, insert)
+        // Generate wrappers and lookup inserts for each tagged function
+        .map(|(f, koto_method_attr)| {
+            let wrapper_name = format_ident!("{PREFIX_FUNCTION}{}", f.sig.ident);
+            let wrapper_fn = wrap_function(f, &wrapper_name, &runtime);
+            let (insert_ops, op_count) =
+                wrapper_function_insert_ops(&f.sig, koto_method_attr, wrapper_name);
+            insert_op_count += op_count;
+            (ImplItem::Fn(wrapper_fn), insert_ops)
         })
         .unzip();
 
-    let item = proc_macro2::TokenStream::from(item);
-    let result = quote! {
-        #item
+    // Append the generated wrapper functions
+    input_struct.items.extend(wrapper_functions);
 
-        #(#wrapper_functions)*
+    // Append entries map initializer and getter functions.
+    // The initializer and getter are separated to avoid "can't use Self from outer item" errors
+    // when dealing with generic impls.
+    let entries_initializer_name = format_ident!("{PREFIX_FUNCTION}initialize_entries_map");
+    let entries_getter_name = format_ident!("{PREFIX_FUNCTION}get_entries_map");
 
-        #[automatically_derived]
-        thread_local! {
-            static #entries_map_name: #runtime::KMap = {
-                let mut result = #runtime::KMap::default();
-                #(#insert_ops)*
-                result
-            };
-        }
+    input_struct
+        .items
+        .push(ImplItem::Fn(koto_entries_initializer(
+            &entries_initializer_name,
+            &insert_ops,
+            insert_op_count,
+            &runtime,
+        )));
 
-        #[automatically_derived]
-        impl #runtime::KotoEntries for #struct_ident {
+    input_struct.items.push(ImplItem::Fn(koto_entries_getter(
+        struct_ident,
+        &entries_getter_name,
+        &entries_initializer_name,
+        &input_struct.generics,
+        &runtime,
+    )));
+
+    // Generate an implementation of KotoEntries that calls the generated entries getter
+    let (impl_generics, ty_generics, where_clause) = input_struct.generics.split_for_impl();
+    let turbofish = ty_generics.as_turbofish();
+    let koto_entries = quote! {
+        impl #impl_generics #runtime::KotoEntries for #struct_type #where_clause {
             fn entries(&self) -> Option<#runtime::KMap> {
-                #entries_map_name.with(|map| Some(map.clone()))
+                Some(#struct_ident #turbofish::#entries_getter_name())
             }
         }
+    };
+
+    let result = quote! {
+        #input_struct
+        #koto_entries
     };
 
     result.into()
 }
 
-fn wrap_function(
-    sig: &Signature,
-    struct_ident: &Ident,
-    runtime: &Path,
-) -> (proc_macro2::TokenStream, Ident) {
-    let type_name = quote! { #struct_ident::type_static() };
-    let fn_name = &sig.ident;
-    let fn_ident = quote! {#struct_ident::#fn_name};
-    let wrapper_name = format_ident!("{PREFIX_FUNCTION}{struct_ident}_{fn_name}");
+fn wrap_function(fn_to_wrap: &ImplItemFn, wrapper_name: &Ident, runtime: &Path) -> ImplItemFn {
+    let fn_name = &fn_to_wrap.sig.ident;
 
-    let arg_count = sig.inputs.len();
-    let mut args = sig.inputs.iter();
+    let arg_count = fn_to_wrap.sig.inputs.len();
+    let mut args = fn_to_wrap.sig.inputs.iter();
 
-    let return_type = detect_return_type(&sig.output);
+    let return_type = detect_return_type(&fn_to_wrap.sig.output);
 
     let wrapper_body = match args.next() {
         // Functions that have a &self or &mut self arg
         Some(FnArg::Receiver(f)) => {
+            // Mutable or immutable instance?
             let (cast, instance) = if f.mutability.is_some() {
                 (quote! {cast_mut}, quote! {mut instance})
             } else {
                 (quote! {cast}, quote! {instance})
             };
 
+            // Does the function expect additional arguments after the instance?
             let (args_match, call_args, error_arm) = match args.next() {
                 None => (
-                    quote! {[]}, // no args expected
-                    quote! {},   // no args to call with
+                    quote! {[]}, // No args expected
+                    quote! {},   // No args to call with
                     quote! { (_, unexpected) =>  #runtime::unexpected_args("||", unexpected)},
                 ),
                 Some(FnArg::Typed(pattern))
                     if arg_count == 2 && matches!(*pattern.ty, Type::Reference(_)) =>
                 {
                     (
-                        // match against any number of args
+                        // Match against any number of args
                         quote! {args},
-                        // append the args to the call
-                        quote! {, args},
-                        // any number of args will be captured
+                        // Append the args to the call
+                        quote! {args},
+                        // Any number of args will be captured
                         quote! { _ => #runtime::runtime_error!(#runtime::ErrorKind::UnexpectedError) },
                     )
                 }
                 _ => panic!("Expected &[KValue] as the extra argument for a Koto method"),
             };
 
-            let call = quote! { #fn_ident(&#instance #call_args) };
-
+            // Wrap the call differently depending on the declared return type
+            let call = quote! { instance.#fn_name(#call_args) };
             let wrapped_call = match return_type {
-                MethodReturnType::None => quote! {
+                MethodReturnType::None => quote! {{
                     #call;
-                    Ok(#runtime::KValue::Null)
-                },
+                    Ok(KValue::Null)
+                }},
                 MethodReturnType::Value => quote! { Ok(#call) },
                 MethodReturnType::Result => call,
             };
 
-            quote! {
-                match ctx.instance_and_args(
-                    |i| matches!(i, #runtime::KValue::Object(_)), #type_name)?
-                {
-                    (#runtime::KValue::Object(o), #args_match) => {
-                        match o.#cast::<#struct_ident>() {
-                            Ok(#instance) => {
-                                #wrapped_call
-                            },
+            quote! {{
+                use #runtime::KValue;
+                match ctx.instance_and_args(|i| matches!(i, KValue::Object(_)), Self::type_static())? {
+                    (KValue::Object(o), #args_match) => {
+                        match o.#cast::<Self>() {
+                            Ok(#instance) => #wrapped_call,
                             Err(e) => Err(e),
                         }
                     },
                     #error_arm,
                 }
-            }
+            }}
         }
+        // Functions that take a MethodContext
         _ => {
-            let call = quote! { #fn_ident(#runtime::MethodContext::new(o, extra_args, ctx.vm)) };
-
+            // Wrap the call differently depending on the declared return type
+            let call = quote! { Self::#fn_name(MethodContext::new(&o, extra_args, ctx.vm)) };
             let wrapped_call = match return_type {
                 MethodReturnType::None => quote! {
                     #call;
-                    Ok(#runtime::KValue::Null)
+                    Ok(KValue::Null)
                 },
                 MethodReturnType::Value => quote! { Ok(#call) },
                 MethodReturnType::Result => call,
             };
 
-            quote! {
+            quote! {{
+                use #runtime::{ErrorKind, KValue, MethodContext, runtime_error};
                 match ctx.instance_and_args(
-                    |i| matches!(i, #runtime::KValue::Object(_)), #type_name)?
+                    |i| matches!(i, KValue::Object(_)), Self::type_static())?
                 {
-                    (#runtime::KValue::Object(o), extra_args) => { #wrapped_call }
-                    _ => #runtime::runtime_error!(#runtime::ErrorKind::UnexpectedError),
+                    (KValue::Object(o), extra_args) => { #wrapped_call }
+                    _ => #runtime::runtime_error!(ErrorKind::UnexpectedError),
                 }
-            }
+            }}
         }
     };
 
-    let wrapper = quote! {
-        #[automatically_derived]
+    let wrapped_fn = quote! {
         fn #wrapper_name(ctx: &mut #runtime::CallContext) -> #runtime::Result<#runtime::KValue> {
             #wrapper_body
         }
     };
 
-    (wrapper, wrapper_name)
-}
-
-fn insert_wrapped_function(
-    sig: &Signature,
-    koto_method_attr: &Attribute,
-    wrapper_name: Ident,
-    runtime: &Path,
-) -> proc_macro2::TokenStream {
-    let fn_name = sig.ident.to_string();
-
-    if matches!(koto_method_attr.meta, Meta::List(_)) {
-        let mut fn_names = vec![fn_name];
-
-        koto_method_attr
-            .parse_nested_meta(|meta| {
-                if meta.path.is_ident("alias") {
-                    let value = meta.value()?;
-                    let s: LitStr = value.parse()?;
-                    fn_names.push(s.value());
-                    Ok(())
-                } else {
-                    Err(meta.error("unsupported attribute"))
-                }
-            })
-            .expect("failed to parse koto_method attribute");
-
-        quote! {
-            let f = #runtime::KValue::NativeFunction(#runtime::KNativeFunction::new(#wrapper_name));
-            #(result.insert(KString::from(#fn_names), f.clone());)*
-        }
-    } else {
-        quote! {
-            result.insert(
-                KString::from(#fn_name),
-                #runtime::KValue::NativeFunction(#runtime::KNativeFunction::new(#wrapper_name)));
-        }
-    }
+    syn::parse2(wrapped_fn).expect("Failed to parse wrapper body")
 }
 
 enum MethodReturnType {
@@ -254,4 +237,171 @@ fn detect_return_type(return_type: &ReturnType) -> MethodReturnType {
             _ => MethodReturnType::Result,
         },
     }
+}
+
+// Generates insertion ops for a wrapped function, checking the attribute for function aliases
+fn wrapper_function_insert_ops(
+    sig: &Signature,
+    koto_method_attr: &Attribute,
+    wrapper_name: Ident,
+) -> (proc_macro2::TokenStream, usize) {
+    let fn_name = sig.ident.to_string();
+
+    if matches!(koto_method_attr.meta, Meta::List(_)) {
+        // Generate additional entries for each function alias
+        let mut fn_names = vec![fn_name];
+
+        koto_method_attr
+            .parse_nested_meta(|meta| {
+                if meta.path.is_ident("alias") {
+                    let value = meta.value()?;
+                    let s: LitStr = value.parse()?;
+                    fn_names.push(s.value());
+                    Ok(())
+                } else {
+                    Err(meta.error("unsupported attribute"))
+                }
+            })
+            .expect("failed to parse koto_method attribute");
+
+        (
+            quote! {
+                let f = KValue::from(KNativeFunction::new(Self::#wrapper_name));
+                #(result.insert(ValueKey::from(#fn_names), f.clone());)*
+            },
+            1 + fn_names.len(),
+        )
+    } else {
+        (
+            quote! {
+                result.insert(
+                    ValueKey::from(#fn_name),
+                    KNativeFunction::new(Self::#wrapper_name).into());
+            },
+            1,
+        )
+    }
+}
+
+fn koto_entries_initializer(
+    entries_initializer_name: &Ident,
+    insert_ops: &[proc_macro2::TokenStream],
+    insert_op_count: usize, // Can be more than insert_ops.len() when aliases are used
+    runtime: &Path,
+) -> ImplItemFn {
+    let initializer_fn = quote! {
+        #[automatically_derived]
+        fn #entries_initializer_name() -> #runtime::KMap {
+            use #runtime::{KMap, KNativeFunction, KValue, ValueKey, ValueMap};
+
+            let mut result = ValueMap::with_capacity(#insert_op_count);
+            #(#insert_ops)*
+            result.into()
+        }
+    };
+
+    syn::parse2(initializer_fn).expect("Failed to parse entries initializer function")
+}
+
+// Generate a thread-safe entries getter
+#[cfg(feature = "arc")]
+fn koto_entries_getter(
+    struct_ident: &Ident,
+    entries_getter_name: &Ident,
+    entries_initializer_name: &Ident,
+    generics: &Generics,
+    runtime: &Path,
+) -> ImplItemFn {
+    let entries_getter = if generics.params.is_empty() {
+        // Non-generic types can cache the entries map in a LazyLock
+        quote! {
+            #[automatically_derived]
+            fn #entries_getter_name() -> #runtime::KMap {
+                use std::sync::LazyLock;
+
+                static ENTRIES: LazyLock<KMap> = LazyLock::new(#struct_ident::#entries_initializer_name);
+
+                ENTRIES.clone()
+            }
+        }
+    } else {
+        // Rust doesn't support generic statics, so entries are cached in a hashmap with the
+        // concrete instantiation type used as the key.
+        let (_impl_generics, type_generics, _where_clause) = generics.split_for_impl();
+        let turbofish = type_generics.as_turbofish();
+
+        quote! {
+            #[automatically_derived]
+            fn #entries_getter_name() -> #runtime::KMap {
+                use std::{any::TypeId, collections::HashMap, hash::BuildHasherDefault, sync::LazyLock};
+                use #runtime::{KCell, KMap, KotoHasher};
+
+                type PerTypeEntriesMap = HashMap<TypeId, KMap, BuildHasherDefault<KotoHasher>>;
+
+                static PER_TYPE_ENTRIES: LazyLock<KCell<PerTypeEntriesMap>> =
+                    LazyLock::new(KCell::default);
+
+                PER_TYPE_ENTRIES
+                    .borrow_mut()
+                    .entry(TypeId::of::<Self>())
+                    .or_insert_with(#struct_ident #turbofish :: #entries_initializer_name)
+                    .clone()
+            }
+        }
+    };
+
+    syn::parse2(entries_getter).expect("Failed to parse entries getter function")
+}
+
+// Generate an entries getter for single-threaded use
+#[cfg(feature = "rc")]
+fn koto_entries_getter(
+    struct_ident: &Ident,
+    entries_getter_name: &Ident,
+    entries_initializer_name: &Ident,
+    generics: &Generics,
+    runtime: &Path,
+) -> ImplItemFn {
+    let entries_getter = if generics.params.is_empty() {
+        // Non-generic types can cache the entries map in a thread-local static
+        quote! {
+            #[automatically_derived]
+            fn #entries_getter_name() -> #runtime::KMap {
+                thread_local! {
+                    static ENTRIES: KMap = #struct_ident::#entries_initializer_name();
+                }
+
+                ENTRIES.with(KMap::clone)
+            }
+        }
+    } else {
+        // Rust doesn't support generic statics, so entries are cached in a hashmap with the
+        // concrete instantiation type used as the key.
+        let (_impl_generics, ty_generics, _where_clause) = generics.split_for_impl();
+        let turbofish = ty_generics.as_turbofish();
+
+        quote! {
+            #[automatically_derived]
+            fn #entries_getter_name() -> #runtime::KMap {
+                use std::{any::TypeId, cell::RefCell, collections::HashMap, hash::BuildHasherDefault};
+                use #runtime::{KMap, KotoHasher};
+
+                type PerTypeEntriesMap = HashMap<TypeId, KMap, BuildHasherDefault<KotoHasher>>;
+
+                thread_local! {
+                    static PER_TYPE_ENTRIES: RefCell<PerTypeEntriesMap> = RefCell::default();
+                }
+
+                PER_TYPE_ENTRIES
+                    .with_borrow_mut(|per_type_entries| {
+                        per_type_entries
+                            .entry(TypeId::of::<Self>())
+                            .or_insert_with(#struct_ident #turbofish :: #entries_initializer_name)
+                            .clone()
+                    })
+            }
+        }
+    };
+
+    syn::parse2(entries_getter).expect("Failed to parse entries getter function")
 }