+++ crates/runtime/src/lib.rs
@@ -23,8 +23,8 @@ pub use crate::{
     types::{
         BinaryOp, CallContext, IsIterable, KCaptureFunction, KFunction, KIterator, KIteratorOutput,
         KList, KMap, KNativeFunction, KNumber, KObject, KRange, KString, KTuple, KValue, KotoCopy,
-        KotoEntries, KotoFunction, KotoHasher, KotoIterator, KotoObject, KotoType, MetaKey,
-        MetaMap, MethodContext, UnaryOp, ValueKey, ValueMap, ValueVec,
+        KotoEntries, KotoField, KotoFunction, KotoHasher, KotoIterator, KotoObject, KotoType,
+        MetaKey, MetaMap, MethodContext, UnaryOp, ValueKey, ValueMap, ValueVec,
     },
     vm::{CallArgs, KotoVm, KotoVmSettings, ModuleImportedCallback, ReturnOrYield},
 };